+++ Modules/Graphics/Direct3D11/src/srk/RenderView.h
@@ -13,9 +13,9 @@ namespace srk::modules::graphics::d3d11 {
 		virtual bool SRK_CALL isCreated() const override;
 		virtual IntrusivePtr<ITextureResource> SRK_CALL getResource() const override;
 		virtual const void* SRK_CALL getNative() const override;
-		virtual uint32_t SRK_CALL getArraySize() const override;
-		virtual uint32_t SRK_CALL getMipSlice() const override;
-		virtual bool SRK_CALL create(ITextureResource* res, uint32_t mipSlice, uint32_t arrayBegin, uint32_t arraySize) override;
+		virtual size_t SRK_CALL getArraySize() const override;
+		virtual size_t SRK_CALL getMipSlice() const override;
+		virtual bool SRK_CALL create(ITextureResource* res, size_t mipSlice, size_t arrayBegin, size_t arraySize) override;
 		virtual void SRK_CALL destroy() override;
 
 		inline ID3D11RenderTargetView1* SRK_CALL getInternalView() {
@@ -23,10 +23,10 @@ namespace srk::modules::graphics::d3d11 {
 		}
 
 	protected:
-		uint32_t _mipSlice;
-		uint32_t _arrayBegin;
-		uint32_t _arraySize;
-		uint32_t _createdArraySize;
+		size_t _mipSlice;
+		size_t _arrayBegin;
+		size_t _arraySize;
+		size_t _createdArraySize;
 
 		IntrusivePtr<ITextureResource> _res;
 		ID3D11RenderTargetView1* _view;
