+++ Modules/Graphics/Direct3D11/src/srk/TextureView.h
@@ -13,9 +13,9 @@ namespace srk::modules::graphics::d3d11 {
 		virtual bool SRK_CALL isCreated() const override;
 		virtual IntrusivePtr<ITextureResource> SRK_CALL getResource() const override;
 		virtual const void* SRK_CALL getNative() const override;
-		virtual uint32_t SRK_CALL getArraySize() const override;
-		virtual uint32_t SRK_CALL getMipLevels() const override;
-		virtual bool SRK_CALL create(ITextureResource* res, uint32_t mipBegin, uint32_t mipLevels, uint32_t arrayBegin, uint32_t arraySize) override;
+		virtual size_t SRK_CALL getArraySize() const override;
+		virtual size_t SRK_CALL getMipLevels() const override;
+		virtual bool SRK_CALL create(ITextureResource* res, size_t mipBegin, size_t mipLevels, size_t arrayBegin, size_t arraySize) override;
 		virtual void SRK_CALL destroy() override;
 
 		inline ID3D11ShaderResourceView* SRK_CALL getInternalView() {
@@ -23,12 +23,12 @@ namespace srk::modules::graphics::d3d11 {
 		}
 
 	protected:
-		uint32_t _mipBegin;
-		uint32_t _mipLevels;
-		uint32_t _createdMipLevels;
-		uint32_t _arrayBegin;
-		uint32_t _arraySize;
-		uint32_t _createdArraySize;
+		size_t _mipBegin;
+		size_t _mipLevels;
+		size_t _createdMipLevels;
+		size_t _arrayBegin;
+		size_t _arraySize;
+		size_t _createdArraySize;
 
 		IntrusivePtr<ITextureResource> _res;
 		ID3D11ShaderResourceView* _view;
