+++ Modules/Graphics/OpenGL/src/srk/BaseTextureView.h
@@ -9,12 +9,12 @@ namespace srk::modules::graphics::gl {
 		~BaseTextureView();
 
 		bool simulative;
-		uint32_t mipBegin;
-		uint32_t mipLevels;
-		uint32_t createdMipLevels;
-		uint32_t arrayBegin;
-		uint32_t arraySize;
-		uint32_t createdArraySize;
+		size_t mipBegin;
+		size_t mipLevels;
+		size_t createdMipLevels;
+		size_t arrayBegin;
+		size_t arraySize;
+		size_t createdArraySize;
 		GLenum internalFormat;
 
 		IntrusivePtr<ITextureResource> res;
