+++ Modules/Graphics/OpenGL/src/srk/RenderViewSimulative.h
@@ -13,9 +13,9 @@ namespace srk::modules::graphics::gl {
 		virtual bool SRK_CALL isCreated() const override;
 		virtual IntrusivePtr<ITextureResource> SRK_CALL getResource() const override;
 		virtual const void* SRK_CALL getNative() const override;
-		virtual uint32_t SRK_CALL getArraySize() const override;
-		virtual uint32_t SRK_CALL getMipSlice() const override;
-		virtual bool SRK_CALL create(ITextureResource* res, uint32_t mipSlice, uint32_t arrayBegin, uint32_t arraySize) override;
+		virtual size_t SRK_CALL getArraySize() const override;
+		virtual size_t SRK_CALL getMipSlice() const override;
+		virtual bool SRK_CALL create(ITextureResource* res, size_t mipSlice, size_t arrayBegin, size_t arraySize) override;
 		virtual void SRK_CALL destroy() override;
 
 	protected:
