+++ Modules/Graphics/OpenGL/src/srk/TextureView.cpp
@@ -23,15 +23,15 @@ namespace srk::modules::graphics::gl {
 		return &_base;
 	}
 
-	uint32_t TextureView::getArraySize() const {
+	size_t TextureView::getArraySize() const {
 		return _base.createdArraySize;
 	}
 
-	uint32_t TextureView::getMipLevels() const {
+	size_t TextureView::getMipLevels() const {
 		return _base.createdMipLevels;
 	}
 
-	bool TextureView::create(ITextureResource* res, uint32_t mipBegin, uint32_t mipLevels, uint32_t arrayBegin, uint32_t arraySize) {
+	bool TextureView::create(ITextureResource* res, size_t mipBegin, size_t mipLevels, size_t arrayBegin, size_t arraySize) {
 		IntrusivePtr guard(res);
 
 		destroy();
